chore(whoop): closeout — address nitpicks + archive OpenSpec change
- whoop/index.ts: drop the per-wave release narration from the package JSDoc.
- labs UI: extract WhoopImportButton so LabImportSection stays under the 60-LOC
  component cap; import the canonical WHOOP_BRIDGE_ID instead of redeclaring it;
  defensive active?.profile?.sex.
- OpenSpec: archive rewrite-whoop-session-bridge (→ archive/2026-07-22-…);
  graduate the whoop-bridge / spa-whoop-extension / whoop-health-adapter
  capability specs and the athlete-connections + health-data updates
  (health-data Purpose now reads nine health types, matching the body).

Completes the WHOOP session-piggyback program.
