+++ openspec/changes/archive/2026-07-22-rewrite-whoop-session-bridge/proposal.md
@@ -1,3 +1,5 @@
+> Completed: 2026-07-22
+
 # Proposal: Rewrite the WHOOP bridge to session piggyback (internal API)
 
 ## Why
