+++ openspec/changes/archive/2026-07-22-rewrite-whoop-session-bridge/specs/whoop-health-adapter/spec.md
