+++ openspec/specs/README.md
@@ -4,7 +4,7 @@ Auto-generated by `scripts/generate-specs-inventory.mjs`. Do NOT hand-edit.
 Run `pnpm specs:inventory` after adding, renaming, or repurposing a spec.
 CI verifies freshness via `pnpm lint:specs-inventory`.
 
-61 specs.
+64 specs.
 
 | Spec | Title | Purpose |
 | ---- | ----- | ------- |
@@ -35,7 +35,7 @@ CI verifies freshness via `pnpm lint:specs-inventory`.
 | [`extension-store-publish`](./extension-store-publish/spec.md) | Extension Store Publish | Generic requirements that any Chrome extension in the monorepo MUST meet before CWS submission: icons, production manifest, store listing, and privacy document… |
 | [`failure-semantics`](./failure-semantics/spec.md) | Failure Semantics | Defines the user- and agent-facing failure contract: the CLI classifies every |
 | [`garmin-bridge`](./garmin-bridge/spec.md) | Garmin Bridge | Chrome extension that proxies Kaiord's workout-editor to Garmin Connect by piggybacking on the user's authenticated session — no credentials stored, CSRF token… |
-| [`health-data`](./health-data/spec.md) | health-data Specification | Defines KRD v2.0 health-domain records — sleep, weight, HRV, daily wellness, body composition, and stress — as first-class `extensions.health.*` payloads with … |
+| [`health-data`](./health-data/spec.md) | health-data Specification | Defines the nine KRD v2.0 health-domain records as first-class `extensions.health.*` payloads with per-metric Zod sub-schemas, round-trip tolerances, and forwa… |
 | [`hexagonal-arch`](./hexagonal-arch/spec.md) | Hexagonal Architecture | Hexagonal (ports-and-adapters) layer structure and dependency rules that every package in the monorepo MUST follow so the domain layer stays pure and adapters … |
 | [`krd-format`](./krd-format/spec.md) | KRD Format | KRD (Kaiord Representation Definition) — the canonical intermediate format through which every conversion (FIT ↔ TCX ↔ ZWO ↔ GCN) must flow. Defines schema, MI… |
 | [`landing-page`](./landing-page/spec.md) | Landing Page | Public landing page at `/` — content, layout, accessibility, performance, and SEO requirements that present Kaiord to first-time visitors and lead into the doc… |
@@ -63,9 +63,12 @@ CI verifies freshness via `pnpm lint:specs-inventory`.
 | [`spa-session-match`](./spa-session-match/spec.md) | spa-session-match Specification | Models the linkage between a planned coaching activity and the executed workout that fulfills it (`SessionMatch`), the application use cases that create / remo… |
 | [`spa-train2go-extension`](./spa-train2go-extension/spec.md) | SPA Train2Go Extension | SPA-side integration with the train2go-bridge extension — runtime discovery via content script announcement and read-only import of coaching plans rendered on … |
 | [`spa-user-preferences`](./spa-user-preferences/spec.md) | spa-user-preferences Specification | Persists per-profile user-facing preferences for the SPA (calendar density today; future fields land here without migration). Reads are lazy and viewport-aware… |
+| [`spa-whoop-extension`](./spa-whoop-extension/spec.md) | spa-whoop-extension Specification | TBD - created by archiving change rewrite-whoop-session-bridge. Update Purpose after archive. |
 | [`spa-workout-state-machine`](./spa-workout-state-machine/spec.md) | SPA Workout State Machine | Workout lifecycle — raw, structured, ready, pushed, modified, stale, skipped — and the transitions (AI generation, explicit finalize, Garmin push, user edits, … |
 | [`test-conventions`](./test-conventions/spec.md) | Test Conventions | The `test-conventions` capability defines two structural invariants every Vitest test SHALL honor: titles begin with `should ` and bodies are organized into `/… |
 | [`test-minimality`](./test-minimality/spec.md) | Test Minimality | Defines the minimal-and-semantic test-suite invariants for every Vitest suite |
 | [`train2go-bridge`](./train2go-bridge/spec.md) | Train2Go Bridge | Chrome extension that imports coaching plans from Train2Go into the workout editor by read-only DOM access on `app.train2go.com` — no credentials stored, no pe… |
 | [`train2go-coach-links`](./train2go-coach-links/spec.md) | train2go-coach-links Specification | End-to-end preservation of Train2Go coach instructions and their embedded links |
 | [`train2go-zones-sync`](./train2go-zones-sync/spec.md) | train2go-zones-sync Specification | Opt-in propagation of athlete threshold and physiological values from Train2Go |
+| [`whoop-bridge`](./whoop-bridge/spec.md) | whoop-bridge Specification | TBD - created by archiving change rewrite-whoop-session-bridge. Update Purpose after archive. |
+| [`whoop-health-adapter`](./whoop-health-adapter/spec.md) | whoop-health-adapter Specification | TBD - created by archiving change rewrite-whoop-session-bridge. Update Purpose after archive. |
