+++ packages/whoop/src/index.ts
@@ -1,19 +1,12 @@
 /**
- * @kaiord/whoop — WHOOP internal-API health adapter for Kaiord.
- *
- * A PURE adapter over the WHOOP internal `core-details-bff` cycles/details,
- * `metrics-service`, and `activities-service` sports catalog responses. It
- * never performs OAuth and never targets the developer API; the SPA injects
- * the read transport. Wave 1 exposes the cycles schema and the
- * recovery→hrv and sleep→sleep converters; Wave 2 adds the read-only
- * cycle→strain and cycle→vitals converters; Wave 3a adds the metrics-service
- * schema and the metrics→heart-rate-series converter; Wave 3b adds the
- * workout schema, the sports catalog, and the workout→activity converter;
- * Wave 4a adds the BFF-tolerant `health-service/v2/stress-bff` schema and
- * the stress-bff→stress-episode converter. Wave 4b adds the
- * `advanced-labs-service` biomarker-tests/summary schemas and
- * `measuredBiomarkers`, filtering the per-test catalog down to what WHOOP
- * actually measured.
+ * @kaiord/whoop — a PURE adapter over the WHOOP internal API (never OAuth,
+ * never the developer API; the SPA injects the read transport). It provides
+ * Zod schemas for the internal responses (`core-details-bff` cycles/details,
+ * `metrics-service`, `activities-service` sports catalog, `stress-bff`,
+ * `advanced-labs-service` biomarkers) and pure converters mapping them to KRD
+ * targets: recovery→hrv, sleep→sleep, cycle→strain, cycle→vitals,
+ * metrics→heart-rate-series, workout→activity, stress-bff→stress-episode, plus
+ * the measured-biomarker filter for Advanced Labs.
  */
 
 // Internal-API response schema & inferred types
